+++ dolfin/io/HDF5File.cpp
@@ -121,9 +121,6 @@ void HDF5File::write(const std::vector<Point>& points,
   // Get number of points (global)
   std::size_t num_points_global = MPI::sum(_mpi_comm, points.size());
 
-  // Data set name
-  const std::string coord_dataset =  dataset_name + "/coordinates";
-
   // Pack data
   const std::size_t n = points.size();
   std::vector<double> x(3*n);
@@ -132,13 +129,12 @@ void HDF5File::write(const std::vector<Point>& points,
       x[3*i + j] = points[i][j];
 
   // Write data to file
-  //  const bool chunking = parameters["chunking"];
   std::vector<std::int64_t> global_size(2);
   global_size[0] = num_points_global;
   global_size[1] = 3;
 
   const bool mpi_io = MPI::size(_mpi_comm) > 1 ? true : false;
-  write_data(coord_dataset, x, global_size, mpi_io);
+  write_data(dataset_name, x, global_size, mpi_io);
 }
 //-----------------------------------------------------------------------------
 void HDF5File::write(const std::vector<double>& values,
@@ -306,8 +302,7 @@ void HDF5File::write(const Mesh& mesh, std::size_t cell_dim,
     std::vector<std::int64_t> topological_data;
     topological_data.reserve(mesh.num_entities(cell_dim)*(num_cell_points));
 
-    const std::vector<std::size_t>& global_vertices
-      = mesh.topology().global_indices(0);
+    const auto& global_vertices = mesh.topology().global_indices(0);
 
     // Permutation to VTK ordering
     const std::vector<std::int8_t> perm = celltype->vtk_mapping();
@@ -367,12 +362,11 @@ void HDF5File::write(const Mesh& mesh, std::size_t cell_dim,
     }
     else
     {
+      // If not already numbered, number entities of order cell_dim
+      DistributedMeshTools::number_entities(mesh, cell_dim);
+
       // Drop duplicate topology for shared entities of less than mesh
       // dimension
-
-      // If not already numbered, number entities of order cell_dim so
-      // we can get shared_entities
-      DistributedMeshTools::number_entities(mesh, cell_dim);
 
       const std::size_t mpi_rank = MPI::rank(_mpi_comm);
       const std::map<std::int32_t, std::set<unsigned int>>& shared_entities
@@ -438,7 +432,10 @@ void HDF5File::write(const Mesh& mesh, std::size_t cell_dim,
     global_size[0] = MPI::sum(_mpi_comm,
                               topological_data.size()/num_cell_points);
     global_size[1] = num_cell_points;
-    dolfin_assert(global_size[0] == (std::int64_t) mesh.size_global(cell_dim));
+
+    const std::int64_t num_cells
+      = mpi_io ? mesh.size_global(cell_dim) : mesh.size(cell_dim);
+    dolfin_assert(global_size[0] == num_cells);
     const bool mpi_io = MPI::size(_mpi_comm) > 1 ? true : false;
     write_data(topology_dataset, topological_data, global_size, mpi_io);
 
@@ -447,9 +444,8 @@ void HDF5File::write(const Mesh& mesh, std::size_t cell_dim,
     {
       const std::string cell_index_dataset = name + "/cell_indices";
       global_size.pop_back();
-      const std::vector<std::size_t>& cell_index_ref
-        = mesh.topology().global_indices(cell_dim);
-      const std::vector<std::size_t> cells(cell_index_ref.begin(),
+      const auto& cell_index_ref = mesh.topology().global_indices(cell_dim);
+      const std::vector<std::int64_t> cells(cell_index_ref.begin(),
             cell_index_ref.begin() + mesh.topology().ghost_offset(cell_dim));
       const bool mpi_io = MPI::size(_mpi_comm) > 1 ? true : false;
       write_data(cell_index_dataset, cells, global_size, mpi_io);
@@ -470,21 +466,6 @@ void HDF5File::write(const Mesh& mesh, std::size_t cell_dim,
     MPI::broadcast(_mpi_comm, partitions);
     HDF5Interface::add_attribute(_hdf5_file_id, topology_dataset,
                                  "partition", partitions);
-
-    // ---------- Markers
-    for (std::size_t d = 0; d <= mesh.domains().max_dim(); d++)
-    {
-      const std::map<std::size_t, std::size_t>& domain
-        = mesh.domains().markers(d);
-      auto _mesh = reference_to_no_delete_pointer(mesh);
-      MeshValueCollection<std::size_t> collection(_mesh, d);
-      std::map<std::size_t, std::size_t>::const_iterator it;
-      for (it = domain.begin(); it != domain.end(); ++it)
-        collection.set_value(it->first, it->second);
-      const std::string marker_dataset
-        = name + "/domain_" + std::to_string(d);
-      write_mesh_value_collection(collection, marker_dataset);
-    }
 
   }
 }
@@ -1261,18 +1242,12 @@ void HDF5File::write_mesh_value_collection(const MeshValueCollection<T>& mesh_va
   global_size[0] = MPI::sum(_mpi_comm, values.size());
   global_size[1] = num_vertices_per_entity;
 
-  // FIXME: this should throw an error, but is here because
-  // "mesh domains" call write_mesh_value_collection with empty
-  // datasets sometimes. Remove when mesh domains are removed.
-  if (global_size[0] > 0)
-  {
-    write_data(name + "/topology", topology, global_size, mpi_io);
+  write_data(name + "/topology", topology, global_size, mpi_io);
 
-    global_size[1] = 1;
-    write_data(name + "/values", value_data, global_size, mpi_io);
-    HDF5Interface::add_attribute(_hdf5_file_id, name, "dimension",
-                                 mesh_values.dim());
-  }
+  global_size[1] = 1;
+  write_data(name + "/values", value_data, global_size, mpi_io);
+  HDF5Interface::add_attribute(_hdf5_file_id, name, "dimension",
+                               mesh_values.dim());
 }
 //-----------------------------------------------------------------------------
 template <typename T>
@@ -1286,7 +1261,7 @@ void HDF5File::write_mesh_value_collection_old(
     = mesh_values.values();
 
   const Mesh& mesh = *mesh_values.mesh();
-  const std::vector<std::size_t>& global_cell_index
+  const std::vector<std::int64_t>& global_cell_index
     = mesh.topology().global_indices(mesh.topology().dim());
 
   std::vector<T> data_values;
@@ -1591,8 +1566,8 @@ void HDF5File::read_mesh_value_collection_old(MeshValueCollection<T>& mesh_vc,
 
     // Get global mapping to restore values
     const Mesh& mesh = *mesh_vc.mesh();
-    const std::vector<std::size_t>& global_cell_index
-      = mesh.topology().global_indices(mesh.topology().dim());
+    const auto& global_cell_index =
+      mesh.topology().global_indices(mesh.topology().dim());
 
     // Reference to actual map of MeshValueCollection
     std::map<std::pair<std::size_t, std::size_t>, T>& mvc_map
@@ -1605,7 +1580,7 @@ void HDF5File::read_mesh_value_collection_old(MeshValueCollection<T>& mesh_vc,
 
     // cells_data in general is not ordered, so we sort it
     // keeping track of the indices
-    std::vector<std::size_t> cells_data_index(cells_data.size());
+    std::vector<std::int64_t> cells_data_index(cells_data.size());
     std::iota(cells_data_index.begin(), cells_data_index.end(), 0);
     std::sort(cells_data_index.begin(), cells_data_index.end(),
               [&cells_data](std::size_t i, size_t j)
@@ -1614,17 +1589,17 @@ void HDF5File::read_mesh_value_collection_old(MeshValueCollection<T>& mesh_vc,
     // The implementation follows std::set_intersection, which we are
     // not able to use here since we need the indices of the
     // intersection, not just the values.
-    std::vector<std::size_t>::const_iterator i = global_cell_index.begin();
-    std::vector<std::size_t>::const_iterator j = cells_data_index.begin();
+    std::vector<std::int64_t>::const_iterator i = global_cell_index.begin();
+    std::vector<std::int64_t>::const_iterator j = cells_data_index.begin();
     while (i!=global_cell_index.end() && j!=cells_data_index.end())
     {
       // Global cell index is less than the cell_data index read from
       // file, if global cell index is larger than the cell_data index
       // read from file, else global cell index is the same as the
       // cell_data index read from file
-      if (*i < cells_data[*j])
+      if (*i < (std::int64_t)cells_data[*j])
         ++i;
-      else if (*i > cells_data[*j])
+      else if (*i > (std::int64_t)cells_data[*j])
         ++j;
       else
       {
@@ -1905,7 +1880,7 @@ void HDF5File::read(Mesh& input_mesh,
 
   // FIXME: explain this more clearly.
   // Reconstruct mesh_name from topology_name - needed for
-  // cell_indices and domains
+  // cell_indices
   std::string mesh_name = topology_path.substr(0, topology_path.rfind("/"));
 
   // FIXME: Imrpove comment - it's unclear this is about
@@ -2017,8 +1992,7 @@ void HDF5File::read(Mesh& input_mesh,
   t.stop();
 
   // Build distributed mesh
-  // FIXME: Why is the mesh built int HDF5Utility? This should be in
-  // the mesh code.
+  // FIXME: Why is the mesh built into HDF5Utility? This should be in the mesh code.
   if (MPI::size(_mpi_comm) == 1)
     HDF5Utility::build_local_mesh(input_mesh, local_mesh_data);
   else
@@ -2027,44 +2001,6 @@ void HDF5File::read(Mesh& input_mesh,
     MeshPartitioning::build_distributed_mesh(input_mesh, local_mesh_data, ghost_mode);
   }
 
-  // ---- Markers ----
-
-  // Check if we have any domains
-  for (std::size_t d = 0; d <= input_mesh.topology().dim(); ++d)
-  {
-    const std::string marker_dataset = mesh_name + "/domain_" + std::to_string(d);
-    if (!has_dataset(marker_dataset))
-      continue;
-
-    auto _mesh = reference_to_no_delete_pointer(input_mesh);
-    MeshValueCollection<std::size_t> mvc(_mesh, d);
-    read_mesh_value_collection(mvc, marker_dataset);
-
-    // Get mesh value collection data
-    const std::map<std::pair<std::size_t, std::size_t>, std::size_t>&
-      values = mvc.values();
-
-    // Get mesh domain data and fill
-    std::map<std::size_t, std::size_t>& markers
-      = input_mesh.domains().markers(d);
-    if (d != input_mesh.topology().dim())
-    {
-      input_mesh.init(d);
-      for (auto entry = values.begin(); entry != values.end(); ++entry)
-      {
-        const Cell cell(input_mesh, entry->first.first);
-        const std::size_t entity_index
-          = cell.entities(d)[entry->first.second];
-        markers[entity_index] = entry->second;
-      }
-    }
-    else
-    {
-      // Special case for cells
-      for (auto entry = values.begin(); entry != values.end(); ++entry)
-        markers[entry->first.first] = entry->second;
-    }
-  }
 }
 //-----------------------------------------------------------------------------
 bool HDF5File::has_dataset(const std::string dataset_name) const
@@ -2084,6 +2020,18 @@ HDF5Attribute HDF5File::attributes(const std::string dataset_name)
   }
 
   return HDF5Attribute(_hdf5_file_id, dataset_name);
+}
+//-----------------------------------------------------------------------------
+void HDF5File::set_mpi_atomicity(bool atomic)
+{
+  dolfin_assert(_hdf5_file_id > 0);
+  HDF5Interface::set_mpi_atomicity(_hdf5_file_id, atomic);
+}
+//-----------------------------------------------------------------------------
+bool HDF5File::get_mpi_atomicity() const
+{
+  dolfin_assert(_hdf5_file_id > 0);
+  return HDF5Interface::get_mpi_atomicity(_hdf5_file_id);
 }
 //-----------------------------------------------------------------------------