+++ dolfin/io/XDMFFile.h
@@ -46,6 +46,7 @@ namespace boost
 namespace pugi
 {
   class xml_node;
+  class xml_document;
 }
 
 namespace dolfin
@@ -79,9 +80,6 @@ namespace dolfin
 
     /// File encoding type
     enum class Encoding {HDF5, ASCII};
-
-    /// Re-use any partition stored in file
-    enum class UseFilePartition : bool {yes=true, no=false};
 
     /// Constructor
     XDMFFile(const std::string filename)
@@ -174,12 +172,49 @@ namespace dolfin
     /// Write out mesh value collection (subset) using an associated
     /// HDF5 file, or storing the data inline as XML.
     ///
-    /// @param    mvc (_MeshValueCollection<std::size_t>_)
+    /// *Arguments*
+    ///     mvc (_MeshValueCollection<bool>_)
+    ///         A list of points to save.
+    ///     encoding (_Encoding_)
+    ///         Encoding to use: HDF5 or ASCII
+    ///
+    void write(const MeshValueCollection<bool>& mvc,
+               Encoding encoding=Encoding::HDF5);
+
+    /// Write out mesh value collection (subset) using an associated
+    /// HDF5 file, or storing the data inline as XML.
+    ///
+    /// *Arguments*
+    ///     mvc (_MeshValueCollection<int>_)
+    ///         A list of points to save.
+    ///     encoding (_Encoding_)
+    ///         Encoding to use: HDF5 or ASCII
+    ///
+    void write(const MeshValueCollection<int>& mvc,
+               Encoding encoding=Encoding::HDF5);
+
+    /// Write out mesh value collection (subset) using an associated
+    /// HDF5 file, or storing the data inline as XML.
+    ///
+    /// *Arguments*
+    ///     mvc (_MeshValueCollection<int>_)
     ///         A list of points to save.
     /// @param    encoding (_Encoding_)
     ///         Encoding to use: HDF5 or ASCII
     ///
     void write(const MeshValueCollection<std::size_t>& mvc,
+               Encoding encoding=Encoding::HDF5);
+
+    /// Write out mesh value collection (subset) using an associated
+    /// HDF5 file, or storing the data inline as XML.
+    ///
+    /// *Arguments*
+    ///     mvc (_MeshValueCollection<double>_)
+    ///         A list of points to save.
+    ///     encoding (_Encoding_)
+    ///         Encoding to use: HDF5 or ASCII
+    ///
+    void write(const MeshValueCollection<double>& mvc,
                Encoding encoding=Encoding::HDF5);
 
     /// Save a cloud of points to file using an associated HDF5 file,
@@ -207,83 +242,104 @@ namespace dolfin
                const std::vector<double>& values,
                Encoding encoding=Encoding::HDF5);
 
-    /// Read in a mesh from the associated HDF5 file, optionally using
-    /// stored partitioning, if possible when the same number of
-    /// processes are being used.
+    /// Read in a mesh
     ///
     /// @param    mesh (_Mesh_)
     ///
-    /// @param    use_file_partition (_UseFilePartition_)
-    ///         Use the existing partition information in HDF5 file
-    ///
-    void read(Mesh& mesh,
-              UseFilePartition use_file_partition=UseFilePartition::no);
+    void read(Mesh& mesh) const;
 
     /// Read first MeshFunction from file
-    ///
-    /// @param meshfunction (MeshFunction)
-    ///   MeshFunction to read in
-    ///
-    void read(MeshFunction<bool>& meshfunction);
+    /// @param meshfunction
+    /// @param name
+    void read(MeshFunction<bool>& meshfunction, std::string name="");
 
     /// Read first MeshFunction from file
-    ///
-    /// @param meshfunction (MeshFunction)
-    ///   MeshFunction to read in
-    ///
-    void read(MeshFunction<int>& meshfunction);
+    /// @param meshfunction
+    /// @param name
+    void read(MeshFunction<int>& meshfunction, std::string name="");
 
     /// Read first MeshFunction from file
-    ///
-    /// @param meshfunction (MeshFunction)
-    ///   MeshFunction to read in
-    ///
-    void read(MeshFunction<std::size_t>& meshfunction);
+    /// @param meshfunction
+    /// @param name
+    void read(MeshFunction<std::size_t>& meshfunction, std::string name="");
 
     /// Read first MeshFunction from file
-    ///
-    /// @param meshfunction (MeshFunction)
-    ///   MeshFunction to read in
-    ///
-    void read(MeshFunction<double>& meshfunction);
+    /// @param meshfunction
+    /// @param name
+    void read(MeshFunction<double>& meshfunction, std::string name="");
 
-    /// Write mesh (new implementation)
-    // FIXME: integrate fully
-    void write_new(const Mesh& mesh, Encoding encoding=Encoding::HDF5) const;
+    /// Read MeshValueCollection from file
+    void read(MeshValueCollection<bool>& mvc, std::string name="");
 
-    /// Read mesh (new implementation)
-    // FIXME: integrate fully
-    void read_new(Mesh& mesh) const;
+    /// Read MeshValueCollection from file
+    void read(MeshValueCollection<int>& mvc, std::string name="");
+
+    /// Read MeshValueCollection from file
+    void read(MeshValueCollection<std::size_t>& mvc, std::string name="");
+
+    /// Read MeshValueCollection from file
+    void read(MeshValueCollection<double>& mvc, std::string name="");
 
   private:
 
+    // Generic MVC writer
+    template <typename T>
+    void write_mesh_value_collection(const MeshValueCollection<T>& mvc,
+                                     Encoding encoding);
+
+    // Generic MVC reader
+    template <typename T>
+    void read_mesh_value_collection(MeshValueCollection<T>& mvc,
+                                    std::string name);
+
+    // Remap meshfunction data, scattering data to appropriate processes
+    template <typename T>
+    static void remap_meshfunction_data(MeshFunction<T>& meshfunction,
+                                        const std::vector<std::int64_t>& topology_data,
+                                        const std::vector<T>& value_data);
+
     // Build mesh (serial)
-    static void build_mesh(Mesh& mesh, std::string cell_type_str,
+    static void build_mesh(Mesh& mesh, const CellType& cell_type,
                            std::int64_t num_points, std::int64_t num_cells,
-                           int num_points_per_cell,
                            int tdim, int gdim,
                            const pugi::xml_node& topology_dataset_node,
                            const pugi::xml_node& geometry_dataset_node,
                            const boost::filesystem::path& parent_path);
 
-   // Build local mesh data structure
-   static void
-   build_local_mesh_data (LocalMeshData& local_mesh_data,
-                          const CellType& cell_type,
-                          const std::int64_t num_points,
-                          const std::int64_t num_cells,
-                          const int num_points_per_cell,
-                          const int tdim, const int gdim,
+    // Build local mesh data structure
+    static void
+      build_local_mesh_data (LocalMeshData& local_mesh_data,
+                             const CellType& cell_type,
+                             std::int64_t num_points, std::int64_t num_cells,
+                             int tdim, int gdim,
+                             const pugi::xml_node& topology_dataset_node,
+                             const pugi::xml_node& geometry_dataset_node,
+                             const boost::filesystem::path& parent_path);
+
+    static void build_mesh_quadratic(Mesh& mesh, const CellType& cell_type,
+                          std::int64_t num_points, std::int64_t num_cells,
+                          int tdim, int gdim,
                           const pugi::xml_node& topology_dataset_node,
                           const pugi::xml_node& geometry_dataset_node,
-                          const boost::filesystem::path& parent_path);
+                              const boost::filesystem::path& relative_path);
+
+
+
+    // Add mesh to XDMF xml_node (usually a Domain or Time Grid) and write data
+    static void add_mesh(MPI_Comm comm, pugi::xml_node& xml_node,
+                         hid_t h5_id, const Mesh& mesh,
+                         const std::string path_prefix);
+
+    // Add set of points to XDMF xml_node and write data
+    static void add_points(MPI_Comm comm, pugi::xml_node& xml_node,
+                           hid_t h5_id, const std::vector<Point>& points);
 
     // Add topology node to xml_node (includes writing data to XML or  HDF5
     // file)
     template<typename T>
     static void add_topology_data(MPI_Comm comm, pugi::xml_node& xml_node,
                                   hid_t h5_id, const std::string path_prefix,
-                                  const Mesh& mesh);
+                                  const Mesh& mesh, int tdim);
 
     // Add geometry node and data to xml_node
     static void add_geometry_data(MPI_Comm comm, pugi::xml_node& xml_node,
@@ -299,16 +355,30 @@ namespace dolfin
                               const std::vector<std::int64_t> dimensions,
                               const std::string number_type="");
 
+    // Calculate set of entities of dimension cell_dim which are duplicated
+    // on other processes and should not be output on this process
+    static std::set<unsigned int> compute_nonlocal_entities(const Mesh& mesh,
+                                                            int cell_dim);
+
     // Return topology data on this process as a flat vector
     template<typename T>
     static std::vector<T> compute_topology_data(const Mesh& mesh, int cell_dim);
 
+    // Return quadratic topology for Mesh of degree 2
+    template<typename T>
+    static std::vector<T> compute_quadratic_topology(const Mesh& mesh);
+
+    // Return data which is local
+    template<typename T>
+    std::vector<T> compute_value_data(const MeshFunction<T>& meshfunction);
+
     // Get DOLFIN cell type string from XML topology node
-    static std::string get_cell_type(const pugi::xml_node& topology_node);
+    static std::pair<std::string, int>
+      get_cell_type(const pugi::xml_node& topology_node);
 
     // Get dimensions from an XML DataSet node
     static std::vector<std::int64_t>
-      get_dataset_shape(const pugi::xml_node& dataset_node);
+    get_dataset_shape(const pugi::xml_node& dataset_node);
 
     // Get number of cells from an XML Topology node
     static std::int64_t get_num_cells(const pugi::xml_node& topology_node);
@@ -320,29 +390,36 @@ namespace dolfin
                                       const boost::filesystem::path& parent_path);
 
     // Return (0) HDF5 filename and (1) path in HDF5 file from a DataItem node
-    std::array<std::string, 2> get_hdf5_paths(const pugi::xml_node& dataitem_node);
+    static std::array<std::string, 2> get_hdf5_paths(const pugi::xml_node& dataitem_node);
 
     static std::string get_hdf5_filename(std::string xdmf_filename);
 
     // Generic MeshFunction reader
     template<typename T>
-    void read_mesh_function(MeshFunction<T>& meshfunction);
+    void read_mesh_function(MeshFunction<T>& meshfunction, std::string name="");
 
     // Generic MeshFunction writer
     template<typename T>
     void write_mesh_function(const MeshFunction<T>& meshfunction,
-                             std::string format, Encoding encoding);
+                             Encoding encoding);
 
-    // Write XML description of point clouds, with value_size = 0, 1
-    // or 3 (for either no point data, scalar, or vector)
-    void write_point_xml(const std::string dataset_name,
-                         const std::size_t num_global_points,
-                         const unsigned int value_size, Encoding encoding);
+    // Get data width - normally the same as u.value_size(), but expand for 2D
+    // vector/tensor because XDMF presents everything as 3D
+    static std::int64_t get_padded_width(const Function& u);
+
+    // Returns true for DG0 Functions
+    static bool has_cell_centred_data(const Function& u);
 
     // Get point data values for linear or quadratic mesh into
-    // flattened 2D array in data_values with given width
-    void get_point_data_values(std::vector<double>& data_values,
-                               std::size_t width, const Function& u) const;
+    // flattened 2D array
+    static std::vector<double> get_point_data_values(const Function& u);
+
+    // Get point data values collocated at P2 geometry points (vertices
+    // and edges) flattened as a 2D array
+    static std::vector<double> get_p2_data_values(const Function& u);
+
+    // Get cell data values as a flattened 2D array
+    static std::vector<double> get_cell_data_values(const Function& u);
 
     // Check whether the requested encoding is supported
     void check_encoding(Encoding encoding) const;
@@ -351,46 +428,6 @@ namespace dolfin
     // enumeration
     static std::string xdmf_format_str(Encoding encoding)
     { return (encoding == XDMFFile::Encoding::HDF5) ? "HDF" : "XML"; }
-
-    // Generate the data string to insert in an xdmf file for the mesh
-    // cell to node connectivity
-    static std::string generate_xdmf_ascii_mesh_topology_data(const Mesh& mesh);
-
-    // Generate the data string to insert in an xdmf file for the mesh
-    // for the topology of entitiy dimension edim -> 0
-    static std::string
-    generate_xdmf_ascii_mesh_topology_data(const Mesh& mesh,
-                                           const std::size_t edim);
-
-    // Generate the data string to insert in an xdmf file for the mesh
-    // point cloud
-    static std::string generate_xdmf_ascii_mesh_geometry_data(const Mesh& mesh);
-
-    // Generate a string of the vertex data numeric values contained
-    // in the data argument which can then be inserted into an xdmf
-    // file. The numeric data is formatted according to the format
-    // argument. E.g:
-    //    generate_xdmf_ascii_data(data_values, "%.15e")
-    template<typename T>
-    static std::string generate_xdmf_ascii_data(const T& data,
-                                                std::string format);
-
-    // As XDMFFile::generate_xdmf_ascii_data, using
-    // boost::lexical_cast to format the numeric data as a string.
-    template<typename T>
-    static std::string generate_xdmf_ascii_data(const T& data);
-
-    // Determine the encoding of the data from the xml file.
-    Encoding get_file_encoding() const;
-
-    // Determine the encoding enumeration value from the xdmf format
-    // string E.g. "XML" or "HDF" See XDMFFile::xdmf_format_str
-    static Encoding get_file_encoding(std::string xdmf_format);
-
-    // Write MVC to ascii string to store in XDMF XML file
-    template <typename T>
-    void write_ascii_mesh_value_collection(const MeshValueCollection<T>& mesh_values,
-                                             std::string data_name);
 
     static std::string vtk_cell_type_str(CellType::Type cell_type, int order);
 
@@ -402,6 +439,9 @@ namespace dolfin
     template <typename T>
     static std::vector<T> string_to_vector(const std::vector<std::string>& x_str);
 
+    // Convert a value_rank to the XDMF string description (Scalar, Vector, Tensor)
+    static std::string rank_to_string(std::size_t value_rank);
+
     // MPI communicator
     MPI_Comm _mpi_comm;
 
@@ -410,22 +450,32 @@ namespace dolfin
     std::unique_ptr<HDF5File> _hdf5_file;
 #endif
 
-    // HDF5 file mode (r/w)
-    std::string _hdf5_filemode;
-
-    // Most recent mesh name
-    std::string _current_mesh_name;
-
     // Cached filename
     const std::string _filename;
 
     // Counter for time series
     std::size_t _counter;
 
-    // The xml document of the XDMF file
-    std::unique_ptr<XDMFxml> _xml;
+    // The XML document currently representing the XDMF
+    // which needs to be kept open for time series etc.
+    std::unique_ptr<pugi::xml_document> _xml_doc;
 
   };
+
+  // Specialisation for std::vector<bool>, as HDF5 does not support it natively
+  template<> inline
+  void XDMFFile::add_data_item(MPI_Comm comm, pugi::xml_node& xml_node,
+                               hid_t h5_id, const std::string h5_path,
+                               const std::vector<bool>& x,
+                               const std::vector<std::int64_t> shape,
+                               const std::string number_type)
+  {
+    // HDF5 cannot accept 'bool' so copy to 'int'
+    std::vector<int> x_int(x.size());
+    for (std::size_t i = 0; i < x.size(); ++i)
+      x_int[i] = (int)x[i];
+    add_data_item(comm, xml_node, h5_id, h5_path, x_int, shape, number_type);
+  }
 
 }